chore: use client test fixture to speedup tests (#81)
This PR introduces a client test fixture. The previous tests
instantiated the client in test class constructors. That is, we used
this pattern:

```csharp
public class MyTests
{
	public MyTests()
	{
		// Instantiate some things
	}

	public void Test_Something()
	{
		// tests
	}

	public void Test_Otherthing()
	{
		// tests
	}
}
```

Data instantiated in the constructor will be setup and torn down
for *each* test method, not just for the class as a whole. This is not
the same thing `setUp` from jUnit.

In Fixtures.cs we create a client fixture and register it to be
available across all test classes. Other tests modified appropriately.