chore: prefer namespace statements to namespace blocks
This PR refactors the code to use namespace statements where
appropriate. In our case, that is everywhere. In C# there are two ways
to place code in a namespace:

(1) Using a namespace block (good when only part of the file should be
in this namespace):
```csharp
namespace MyNamespace
{
	// Code here indented +1 level
}
```

(2) To the entire file using a statement
```csharp
namespace MyNamespace;

// Code here with no extra indent necessary
```

All of our code used style (1), which was unncessary. In every case
the namespace applied to the whole file. Because we opted to use
blocks, every file carried an extra level of indentation.

This PR undoes this by switching to namespace statements. We unindent
each file by one level as a result.