Create custom comprarer to handle containers byte arrays.
In C# byte array comparisons are by reference. This will likely cause
unexpected behavior to end-users. To make things do what we mean, we
provide a structural comprarer that works on the contents as opposed
to the array reference.

We also provide a `DictionaryEquals` extension that parallels the C#
library's `SetEquals` method. We have to do this because, even if we
provide a structural comparer as described above, this only operates
on the keys, not the values.