Use `ToStringUtf8` to reduce the number of conversions.
Protobuf provides methods `ToByteArray` and `ToStringUtf8`. This
commit changes these GetResponse classes to store the protobuf
ByteStrings, then convert either to byte array or directly to string.

Previously to convert to a string we had to convert to byte array and
then to string. This removes the intermediate to byte array conversion.