feat: list push, pop, and fetch (#115)
* Bump protos version.

* Implement ListPushFront and ListFetch. Test ListFetch.

* Test `ListPushFront`.

* Implement and test `ListPushFrontBatch`.

In this commit we also add more extension methods to convert
`IEnumerable<byte[]>` and `IEnumerable<string>` to
`IEnumerable<ByteArray>`.

* Remove `ListPushFrontBatch` methods.

We will expose a `Concatenate` API to serve this need.

* Add `ListPushFrontAsync` tests to exercise push happy path semantics

* Implement `ListPushBackAsync` and test.

* Implement and test `ListPopFront`.

* Improve `ListPopFront` test to disambiguate popping front vs back

* Implement and test `ListPopBack`.

* Refactor unary get responses.

Many data structures have `Get` operations that retrieve a unary
value with a status. The data and accessors are exactly the same as
for the primitive `CacheGetStatus`. To reduce code duplication, we
refactor the common data and behavior into a base class and derive
from that.

* Document list methods.

* Bump to latest protos.