chore: remove methods with variadic args
This PR removes methods with variadic args. These overloads are
undesirable for a few reasons:
(1) They add considerable surface area to SDK.

(2) Because the arguments (cache keys/values) are variable length,
they go at the end of the method signature. This breaks the pattern
where the cache keys/values go after the cache name.

(3) Mixed support across languages. We want to add the most common
case first. In most languages, the common case is not variable length
arguments as opposed to lists.