fix: make truncate*ToSize nomenclature consistent with `ListPush*` (#126)
* fix: make truncate*ToSize nomenclature consistent with `ListPush*`

For `ListPush` We use the nouns "Front" and "Back" to refer to the
beginning and end of the list respectively. In the previous commit, we
used the nouns "Head" and "Tail". That introduced extra vocabulary to
refer to exactly the same concepts. This commit renames "Head" and
"Tail" to "Front" and "Back" for consistency.

* doc: clarify front/back in documentation.