fix: bad reference to error email action
The CI job uses a github action that sends e-mails to us on build failures.
That action was defined in the examples repo, and when I moved
the dotnet examples over to this repo I didn't bring that action
along, so there was a bad reference to it in the ci.yaml.

Now that action has been moved to standards-and-practices repo
so it can be shared across other projects.  This commit fixes
the reference to point to the standards-and-practices repo.