feat: Initial sketch of middleware implementation
This commit introduces a proposed implementation for our Middleware.

It is an attempt to expose the most relevant parts of the gRPC
Interceptor in a way that is a bit more approachable for end users
who do not want to become gRPC experts.  It still (by necessity)
exposes some of the gRPC constructs (such as the Metadata container
that stores headers, and the gRPC status code), but it abstracts
away some of the more obtuse constructs in the interceptor
(such as AsyncUnaryCall).

As implemented, this only works for AsyncUnary requests, but we
should not currently be making any other types of requests on
the data plane.  If we need to support blocking requests (doubtful)
or Streaming requests (more plausible), and we want to support
middleware on those, we will need to extend the API.

In subsequent commits this can be used to implement retries, and
to limit the maximum number of concurrent requests.  In the future
it can be used for logging client-side performance data.