feat: add logger to client (#181)
This PR allows the cache client to accept a `ILoggerFactory`
parameter. This is used to create loggers in the cache client itself,
any contained objects, and so on.

We allow the `loggerFactory` to be `null`. In this case we use the
`NullLoggerFactory` singleton to create concrete loggers. Those
loggers will just perform no-ops.