feat: max concurrent request middleware (#182)
During tuning, I discovered that the .NET grpc library behaves
differently than the others I'd previously tested w/rt connections
opened to the server.  Specifically, it will detect whenever your
number of concurrent requests exceeds a multiple of 100 and open
a new connection to the server when it does.  Thus, in my loadgen
scenario where I was executing 5000 concurrent requests, the library
would end up opening on the order of 50 connections to the server.

This behavior is undesirable for both the client and server, depending
on the network bandwidth available between the two.  Specifically, in
a laptop environment, throughput and latency both decrease when you
exceed about 200 concurrent requests, so there is no value in opening
so many additional connections to the server and this could cause
server-side throttling/load shedding.

In this commit we introduce a new middleware that can restrict
the maximum number of concurrent requests.  Setting this to 200
in the laptop environment results in far fewer connections, and
throughput and latency for the happy path cases are unaffected.
For the degenerate case (5000+ concurrent requests), this protects
the server and actually seems to improve client-side p999 latencies
by quite a bit.

Note that as part of this commit I needed to re-work the middleware
layer.  We can't rely directly on using gRPC Interceptors for this,
because their method signatures are not asynchronous.  Thus, we
would end up blocking threads to wait for any kind of semaphore,
which would have a big negative impact on performance.  Instead, in
this commit, we move the middleware layer to some code that we own
ourselves so that we are able to use async method signatures.