chore: wire up grpc settings from Configuration
This commit updates our interface for the gRPC settings.
The original version was built based on the specification
doc that I wrote, which was in turn based on tuning work
done with our JavaScript SDK.  In .NET it turns out that
the settings we were configuring for JS aren't relevant,
and furthermore the .NET gRPC library only exposes a much
more strongly-typed interface into what gRPC channel options
you can configure.  This is largely what we were trying
to accomplish with our abstraction, so it makes more sense
for this SDK to just expose the .NET one directly, which
is what we do in this commit.