fix: remove unnecessary deps (e.g. duplicate grpc dep) (#263)
* fix: remove unnecessary deps (e.g. duplicate grpc dep)

This commit attempts to pare down our dependency list to the
minimum requirements to compile.  Previously we had more than
one gRPC library, plus a google protobuf lib that we don't seem
to need.

This also removes the SourceLink.Github because it wasn't required
for compiling, but that is probably a mistake and should be
reverted... just posting this for discussion since I'm not familiar
with that one.