fix: handle AggregateExceptions from task continuations
In Middleware code, if you need to modify the response before it
makes it back out to the caller, you need to use `Task.ContinueWith`.
However, if an exception is thrown while such a Middleware is
in the stack, the exception will get wrapped in an AggregateException.

Prior to this commit our CacheExceptionMapper would treat this as
an unknown error and we would lose our desired status code mapping.
This commit unwraps AggregateExceptions before the mapping so that
the error codes will be correct.

It also adds debug logs to the exception mapper so that we can
see what's going wrong if other exception types fail to map
properly in the future.