feat: add basic retry middleware
This commit adds an initial implementation of a retry middleware,
along with fleshing out the FixedCountRetryStrategy for compatibility
with the new middleware.  As of this commit, any Get/Set requests
that fail with an Unavailable or Internal error gRPC status code
will be retried up to 3 attempts.

There is no backoff or jitter yet; we should implement that in the
future.

The determination of which gRPC status codes and request types are
eligible for retries is implemented via a new IRetryEligibilityStrategy
interface.  This way, individual retry strategies have the ability
to override the behavior, but in the 90% case they can just re-use
the default eligibility strategy.

NOTE: we implement retries as a bespoke middleware rather than using
the retry configuration that is built into the .NET gRPC library
because the built-in implementation has some strict restrictions
that aren't compatible with our goals here (e.g. no response that
includes any headers is eligible for retry in the built-in implementation,
and we always get headers back from envoy.)