feat: prefer `int` to `uint`
This updates the code across the board to prefer the type `int` to
`uint`. The latter type is discouraged in .NET, as it is not broadly
compatible with other .NET languages. See here for more details:
https://learn.microsoft.com/en-us/dotnet/standard/language-independence

Not only is it discouraged, but the various date time classes in .NET
prefer `int` to `uint` as well. By preferring `int`, we avoid forcing
end-users to library values to `uint`.

Lastly we add checks to the API that TTLs are strictly positive when
provided.