feat: simplify logging configuration
Kenny called out on a previous PR that there was some code smell
in the way I was managing the logging factories for the various
components of the configuration.

This commit reworks things slightly so that the user still
only has to pass in the ILoggerFactory in one spot in our
pre-built configs, but we now just pass that directly through
to the constructors of the other things that need it.  This
allows us to get rid of the getters and copy constructors
that we had previously exposed.