feat: use `TimeSpan` to specify TTL and deadline (#286)
This updates the code across the board to use TimeSpan instead of uint (milli)?seconds. TimeSpan is more ergonomic and used broadly in other client APIs. The latter type is discouraged in .NET, as it is not broadly compatible with other .NET languages. See here for more details: https://learn.microsoft.com/en-us/dotnet/standard/language-independence