feat: Rename `Exception` to `InnerException`
This commit changes the name of the `Exception` object on our
Error classes to `InnerException`.  This seems like the pattern
that is used in the existing C# exception object model for
cases like ours, where you don't usually expect the caller to
dig into the nested error details but you want to make it possible
for them to do so.