chore: hopefully final touches to README
Tweaks the ordering of the README sections a bit to make the
examples more prominent and highlight the pattern matching
earlier.

Also adds an initial doc for advanced configuration.

The README template generator will need to be tweaked to
accommodate these changes but this feels like a better
structure based on the feedback we got in the bug bash.