chore: migrate EagerConnection to EagerConnectionTimeout
This commit changes the new EagerConnection setting from a boolean
to a TimeSpan.  This allows the user to specify a timeout for how
long they wish to wait for the eager connection to be established
at construction time, so that they can time out more aggressively
rather than blocking at startup if they prefer.