feat: add support for configuring eager connection on client (#348)
* feat: add support for configuring eager connection on client

This commit adds a configuration setting that can be used to
enable eager connection to the server when the client is
constructed.  By default, the client establishes the connection
lazily when the first request is executed.  However, this
approach can lead to request timeouts on the first request
if the connection takes too long to establish, because the
connection time is counted toward the client-side timeout.
When eager connections are enabled, the client will attempt
to establish the connection to the server at construction time,
which means that the connection will already be established before
the first user request is issued.