feat: support configuring a minimum number of gRPC channels (#367)
This commit adds a setting to gRPC config that allows users to
configure a minimum number of gRPC channels.  Prior to this,
the client would always construct a single channel and would
dynamically add additional ones only after we exceeded the
maximum number of concurrent requests on the current channels.
This results in imbalanced load on the server for cases where
the number of concurrent requests remains below the threshold.

Users with this type of workload can increase this new setting
to ensure they have more than one connection to the servers,
and thus increase the likelihood of balancing requests
across multiple servers for improved performance.