feat: Change name of SimpleCacheClient to CacheClient
As per feedback from dev advocates and users, the name SimpleCacheClient
is a little confusing and we prefer CacheClient.  This commit
implements the name change.

NOTE: we need to retain a deprecated alias using the old name,
so that we don't break backward compatibility.  That will be done
in a separate commit because the diff is too complex to review
when it is done in the same commit.