feat: support .NET Framework 4.61+ using gRPC-Web (#426)
Because .NET Framework does not support gRPC over HTTP/2, we fall back
to using gRPC for those older versions. To do this we:

Add another target framework for the library, net461. This is the
lowest .NET Framework version also supported by netstandard2.0.

Conditional on a net461 build, link the gRPC-Web client
dependency.

Conditional on a net461 build, use a gRPC-Web HttpHandler for
the control and data gRPC channels.

Update the test projects to exercise the net461 build as well.