chore: add interface for all error response types
In the current formulation, downstream code cannot consume a generic
`Error` type, because current error types inherit from a parent
abstract class.

This PR introduces an error interface, `IError`, which specifies that
an error response has the following properties:
- `InnerException`
- `ErrorCode`
- `Message`

which is common to all error classes.

With this in place we inherit the documentation from the interface,
simplifying the doc strings in the response classes.