chore: get rid of TEST_CACHE environment variable
Our tests currently require a TEST_CACHE environment variable, which
is not really useful and sets a pattern that other SDKs may need
to require this too in order to standardize on a way to launch them
(for example in the SDK canaries). The JS SDK does not work properly
if you set that environment variable, because it is intended for use
when running against a local mr2.

This commit removes the env var and dynamically creates a cache
name to use for the tests.