chore: Relax integration test timing assertions (#465)
Momento doesn't offer strict read-after-write consistency, despite
often appearing to. That's an implicit test dependency on the
subjective latency of the test runner, however. By relaxing the
expectation to some small window for a get() response to settle,
the tests should be a little more reliable.

I wrapped CacheClient in a new TestCacheClient implementation of
ICacheClient. Code for delegation to the TestCacheClient's client
field was generated by Visual Studio. This new implementation is
the new CacheClientFixture client implementation.

Also minorly updated some docs.