feat: Add build for Unity platform
This commit makes a small change to the csproj file that allows us
to specify DefineConstants=USE_GRPC_WEB for any target framework.
That allows us to do a netframework20 + grpc-web build, which is
necessary for compatibility with the Unity platform.

We also add a step to the release action to do a build with this
configuration, and attach it to the github release.