feat: add topics (#468)
Add the initial topics implementation. The entry point is TopicClient,
which contains the publish and subscribe methods. Subscribe returns an
IAsyncEnumerable<TopicMessage> that can be iterated over to read from
the topic.

TopicMessage is either a Text, a Binary, or an Error.

Add new configurations for creating a topic client.

---------

Co-authored-by: Kenny <kenny@momentohq.com>