feat: implement mvi count items command
Adds MVI count items, which returns the number of items in a vector
index. In the event an index does not exist, the response is a
`NOT_FOUND` error. That way we distinguish between an empty index that
does exist vs an index that does not exist.

This PR adds the response type, method, documentation, and integration
tests.