fix: include exception in log message when eager connection fails; update env vars to MOMENTO_API_KEY
Prior to this commit, if we failed to eagerly connect, we were not logging any information about
the reason. This commit adds the exception to the log. Tested by disabling my network connection
and verified it includes all the relevant info.

Also updates places in example code where we were still using MOMENTO_AUTH_TOKEN for env var names,
now that we have standardized on MOMENTO_API_KEY.