feat: bump minimun NET Framework from 461 to 462
Previously we relied on .NET Framework 4.61 as the minimum
version. The next minor version introduced async iterator features
necessary for the topic client. Because of this, at the time, we added
preprocessor directives to exclude the topic client if the runtime
wasn't .NET Standard 2.0 or greater.

Now that we add another target framework, .NET 6.0, maintaining the
different platform directives is difficult. Instead we bump the target
framework from net461 to net462 to gain the async iterators and remove
the platform conditional compilation directives.