chore: remove socket option modifier from configuration
Previously we had added a copy factory to `Configuration` to change
`SocketHttpHandlerOptions`. While tihs was for the convenience of the
`Configurations` code, this exposes a low level option to the top
level of `Configuration`, which is undesirable.

We remove that in this PR. The previous code using it was wholly
encapsulated in `Configurations`, so not a breaking change.