chore: refactor makefile and actions (#565)
Simplifies the Makefile in the following ways:

Refactors the core commands build and test to be OS-conditional. When we detect Windows, we build and test against not just .NET 6.0 but also .NET Framework, the latter of which is Windows-only. This way a developer or consumer of the Makefile does not need to specify an OS-specific target (test-dotnet-framework vs test-dotnet-6)
Adds a make variable for building with gRPC web. The GitHub Actions tested against each OS, Framework, and with/without gRPC web. We make this easier by adding a GRPC_WEB variable. To build with gRPC web, run make GRPC_WEB=true build.
Includes logging options previously only in GitHub actions.
Refactors GitHub Actions to invoke the Makefile targets to build and test. When on Windows, build builds both .NET 6.0 and .NET Framework 4.62 targets; the same goes for when testing.
Moves PHONY declarations to the top for readability.