feat: Add a separate web SDK for clients without HTTP 2.0
Add a new Momento.Sdk.Web artifact that uses gRPC-Web so that users
without access to HTTP 2.0 can use the SDK. Building with the
USE_GRPC_WEB will generate the new Momento.Sdk.Web.dll. Note that this
will affect the unity build as well.

Add a pack and publish step to push the new artifact to nuget.