chore: Add a Web SDK example
Add an example that uses the .NET Web SDK. It is a copy of the basic
application example.

Add a step to the github build that runs the web example to ensure we
test the web SDK.

Update the examples to the latest SDK version.

Change the eager connection example to use the CreateAsync function to
make it compatible with the newest versions of the SDK.

Misc fixes.