feat: introduce `ITopicEvent` to cover messages and system events
Introduces an interface `ITopicEvent` which covers both system
events (`TopicEvent`) and normal messages (`TopicMessage`). We
refactor the internals of the stream to be over `ITopicEvent`s, so
that we can filter to just `TopicMessage` (as is implemented
currently) or pass through all events in an event enumerable (to come
in a future commit).