feat: add an topic event enumerable (#575)
Introduces an interface ITopicEvent which covers both normal
messages (TopicMessage) and system events (TopicSystemEvent). We
refactor the internals of the stream to be over ITopicEvents, so
that we can filter to just TopicMessage (as is implemented
currently) or pass through all events in an event enumerable.

To enumerate over all events in the stream, call GetAllEventsAsyncEnumerable,
or WithCancellationForAllEvents.

This largely parallels the existing enumerator over messages only. See the
integration tests for an example usage.