feat: Add the read concern header
Add the read concern header to the top level Configuration.

Set the consistent read concern header in the cache client test fixture
if the CONSISTENT_READS env var is set.

Add a new make target for the github build to use that sets the header.

Update the test-cache-service make target to set the header so that the
canaries will use consistent reads.